+++ db/seeds.rb
@@ -56,22 +56,35 @@ jr = FactoryBot.create(:resident, last_name: 'Rainwater', first_name: 'Jim', ema
     verified_on: 1.day.ago)
 end
 
-lot24 = FactoryBot.create(:lot, lot_number: '24 (T)', size: 1, paid_on: 1.day.ago)
-lot42 = FactoryBot.create(:lot, lot_number: '42 (T)', size: 0.5)
-property_123Main = FactoryBot.create(:property, lots: [lot24], street_number: '123', street_name: 'Main St (TEST)')
-property_975Main = FactoryBot.create(:property, lots: [lot42], street_number: '975', street_name: 'Main St (TEST)')
+lot11 = FactoryBot.create(:lot, lot_number: '11 (T)', size: 1, paid_on: 1.day.ago)
+lot12 = FactoryBot.create(:lot, lot_number: '12 (T)', size: 0.5)
+property_123Main = FactoryBot.create(:property, lots: [lot11], street_number: '123', street_name: 'Main St (TEST)')
+property_975Main = FactoryBot.create(:property, lots: [lot12], street_number: '975', street_name: 'Main St (TEST)')
 
 jqo = FactoryBot.create(:resident, last_name: 'Owner', first_name: 'Jane', email_address: 'janeowner@example.com')
 
 FactoryBot.create(:residency, property: property_123Main, resident: jqo, resident_status: :deed_holder, verified_on: 1.day.ago)
 FactoryBot.create(:residency, property: property_975Main, resident: jqo, resident_status: :deed_holder, verified_on: 1.day.ago)
 
-jqr = FactoryBot.create(:resident, last_name: 'Renter', first_name: 'John', email_address: 'johnqrenter@example.com', properties: [property_975Main]).tap do |resident|
+jqr = FactoryBot.create(:resident, last_name: 'Renter', first_name: 'John (no email)', email_address: nil, properties: [property_975Main]).tap do |resident|
   resident.residencies.first.update(
     resident_status: :renter,
     verified_on: 1.day.ago
   )
 end
+
+bqr = FactoryBot.create(:resident, last_name: 'Renter', first_name: 'Bob', email_address: 'bobqrenter@example.com', properties: [property_975Main]).tap do |resident|
+  resident.residencies.first.update(
+    resident_status: :renter,
+  )
+end
+
+jdoe = FactoryBot.create(:resident, last_name: 'Doe', first_name: nil, email_address: nil, properties: [property_975Main]).tap do |resident|
+  resident.residencies.first.update(
+    resident_status: :renter,
+  )
+end
+
 
 # Comments
 FactoryBot.create(:comment, commentable: mms)